fixes #5755
background: for the portal to define a flatvocab that could be enum in a components site, we had it in wdkqueryplugin that if the VQ does not exist in the componnet site, then it is an enum.

But this broke genes by location,when DD strated using generic param "organism" stating a specific queryRef "VQ.withChromosomes".
The param "organism" (flatvocab) goes to all sites, exists in all sites,  and  when this VQ.withChromosomes does not exist (for exmaple: Trich, the comp site uses teh default defined in its flatvocab : VQ.withGenes..... and the portal ends up with ALL organisms when only  those with Chromosomes are needed.

The fix here changes the logic so when the param requested is a flatvocab in the component site, if the VQ does not exist, it does nothing.




git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonWebService/trunk@42532 735e2a04-f8fc-0310-8a1b-f2942603c481